AI_Model/requirements.txt
# Hand-Pet Interaction Detector - Requirements

# Deep Learning Frameworks
torch>=1.10.0
torchvision>=0.11.0

# Computer Vision
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0

# YOLO Models (choose one or both)
# For YOLOv5
# git+https://github.com/ultralytics/yolov5

# For YOLOv8
ultralytics>=8.0.0

# Optional: MediaPipe for lightweight hand detection
mediapipe>=0.8.11

# Utilities
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.3.0
pillow>=8.0.0
pyyaml>=5.4.0
tqdm>=4.62.0

# Data annotation (optional)
labelImg>=1.8.6

# For Jetson Nano optimization
# onnx>=1.10.0
# onnxruntime-gpu>=1.10.0

# Logging and monitoring
tensorboard>=2.7.0

# Alert system
# winsound (built-in for Windows)

# Testing
pytest>=6.2.0
